# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cross App Backend",
    "description": "API para gestao e gamificacao de CrossFit",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desenvolvimento local"
    }
  ],
...
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login de usuario",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "JWT de identidade retornado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Credenciais invalidas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/boxes/register": {
      "post": {
        "tags": [
          "Boxes"
        ],
        "summary": "Cadastro do contratante",
        "description": "Cria um novo Box (ou filial via parentBoxId), cria ou vincula o ADMIN informado e ja retorna o JWT da sessao autenticada para o novo box.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterBoxDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Box cadastrado com sucesso e JWT retornado automaticamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterBoxResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados invalidos ou duplicados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/boxes/mine": {
      "get": {
        "tags": [
          "Boxes"
        ],
        "summary": "Lista os boxes do usuario autenticado",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de boxes retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Box"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, invalido ou expirado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/checkins": {
      "post": {
        "tags": [
          "Checkins"
        ],
        "summary": "Realiza check-in por geolocalizacao no box",
        "description": "Permite check-in apenas quando o atleta estiver a menos de geofenceRadius metros das coordenadas do box e ja estiver matriculado nele.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/XBoxIdHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCheckinDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Check-in realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckinCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados invalidos para realizacao de check-in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Usuario ainda nao matriculado no box ou fora do raio permitido para check-in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Box nao encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/checkins/me": {
      "get": {
        "tags": [
          "Checkins"
        ],
        "summary": "Lista os checkins do usuario autenticado",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Checkins do usuario retornados com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Checkin"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/checkins/box": {
      "get": {
        "tags": [
          "Checkins"
        ],
        "summary": "Relatorio diario de checkins do box",
        "description": "Retorna todos os checkins registrados no box atual no dia de hoje.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/XBoxIdHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "Checkins do dia retornados com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Checkin"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Perfil sem permissao para acessar relatorio",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/exercises": {
      "get": {
        "tags": [
          "Exercises"
        ],
        "summary": "Lista exercicios do box + exercicios globais",
        "description": "Retorna todos os exercicios globais da plataforma (isGlobal: true) mais os exercicios customizados do box do usuario autenticado.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/XBoxIdHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de exercicios retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Exercise"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, invalido ou expirado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Exercises"
        ],
        "summary": "Cria exercicio customizado no box",
        "description": "Insere um novo exercicio vinculado ao box do Admin autenticado. Acesso restrito a ADMIN.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/XBoxIdHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExerciseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Exercicio criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateExerciseResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados invalidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, invalido ou expirado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Perfil sem permissao para criar exercicio",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Exercicio com este nome ja existe neste box",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Cadastro independente de aluno",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario cadastrado e JWT retornado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "409": {
            "description": "Ja existe usuario com este email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/boxes/nearby": {
      "get": {
        "tags": [
          "Boxes"
        ],
        "summary": "Lista academias proximas com base na latitude e longitude do usuario",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "latitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "minimum": -90,
              "maximum": 90,
              "example": -23.56447
            },
            "description": "Latitude atual do usuario"
          },
          {
            "name": "longitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "minimum": -180,
              "maximum": 180,
              "example": -46.65284
            },
            "description": "Longitude atual do usuario"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de academias proximas retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NearbyBox"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, invalido ou expirado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "XBoxIdHeader": {
        "name": "x-box-id",
        "in": "header",
        "required": true,
        "description": "ID do box selecionado",
        "schema": {
          "type": "string",
          "example": "67ea76a5ac5d89c8bb9d2111"
        }
      }
    },
    "schemas": {
      "ObjectId": {
        "type": "string",
        "pattern": "^[a-fA-F0-9]{24}$",
        "example": "67ea76a5ac5d89c8bb9d2111"
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "ADMIN",
          "COACH",
          "ALUNO"
        ]
      },
      "ExerciseCategory": {
        "type": "string",
        "enum": [
          "WEIGHTLIFTING",
          "GYMNASTICS",
          "MONOSTRUCTURAL",
          "ACCESSORY"
        ]
      },
...
      "LoginDto": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "admin@box.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "Senha@123"
          }
        }
      },
      "RegisterBoxDto": {
        "type": "object",
        "required": [
          "name",
          "cnpj",
          "latitude",
          "longitude",
          "geofenceRadius",
          "adminName",
          "adminEmail",
          "adminPassword"
        ],
        "properties": {
          "parentBoxId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120,
            "example": "Cross Box Alpha"
          },
          "cnpj": {
            "type": "string",
            "pattern": "^\\d{14}$",
            "example": "12345678000199"
          },
          "latitude": {
            "type": "number",
            "minimum": -90,
            "maximum": 90,
            "example": -23.56447
          },
          "longitude": {
            "type": "number",
            "minimum": -180,
            "maximum": 180,
            "example": -46.65284
          },
          "geofenceRadius": {
            "type": "number",
            "minimum": 1,
            "maximum": 5000,
            "example": 100
          },
          "adminName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120,
            "example": "Ana Admin"
          },
          "adminEmail": {
            "type": "string",
            "format": "email",
            "example": "admin@box.com"
          },
          "adminPassword": {
            "type": "string",
            "minLength": 8,
            "example": "Senha@123"
          }
        }
      },
      "CreateCheckinDto": {
        "type": "object",
        "required": [
          "latitude",
          "longitude"
        ],
        "properties": {
          "latitude": {
            "type": "number",
            "minimum": -90,
            "maximum": 90,
            "example": -23.56447
          },
          "longitude": {
            "type": "number",
            "minimum": -180,
            "maximum": 180,
            "example": -46.65284
          }
        }
      },
      "CreateExerciseDto": {
        "type": "object",
        "required": [
          "name",
          "category"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "example": "Snatch"
          },
          "category": {
            "$ref": "#/components/schemas/ExerciseCategory"
          }
        }
      },
...
      "JwtUser": {
        "type": "object",
        "required": [
          "sub",
          "email",
          "boxIds",
          "role"
        ],
        "properties": {
          "sub": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "boxIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        }
      },
      "SessionResponse": {
        "type": "object",
        "required": [
          "accessToken",
          "tokenType",
          "user"
        ],
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "tokenType": {
            "type": "string",
            "example": "Bearer"
          },
          "user": {
            "$ref": "#/components/schemas/JwtUser"
          }
        }
      },
      "RegisterBoxResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SessionResponse"
          },
          {
            "type": "object",
            "required": [
              "boxId",
              "adminId",
              "adminLinkedToExisting",
              "message"
            ],
            "properties": {
              "boxId": {
                "$ref": "#/components/schemas/ObjectId"
              },
              "adminId": {
                "$ref": "#/components/schemas/ObjectId"
              },
              "adminLinkedToExisting": {
                "type": "boolean",
                "example": false
              },
              "message": {
                "type": "string",
                "example": "Box, usuario ADMIN e JWT criados com sucesso"
              }
            }
          }
        ]
      },
      "BoxLocation": {
        "type": "object",
        "required": [
          "type",
          "coordinates"
        ],
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "Point"
            ]
          },
          "coordinates": {
            "type": "array",
            "description": "Formato [longitude, latitude]",
            "minItems": 2,
            "maxItems": 2,
            "items": {
              "type": "number"
            },
            "example": [
              -46.65284,
              -23.56447
            ]
          }
        }
      },
      "Box": {
        "type": "object",
        "required": [
          "_id",
          "name",
          "cnpj",
          "location",
          "geofenceRadius",
          "createdAt"
        ],
        "properties": {
          "_id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "parentBoxId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "name": {
            "type": "string",
            "example": "Cross Box Alpha"
          },
          "cnpj": {
            "type": "string",
            "example": "12345678000199"
          },
          "location": {
            "$ref": "#/components/schemas/BoxLocation"
          },
          "geofenceRadius": {
            "type": "number",
            "example": 100
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CheckinCreateResponse": {
        "type": "object",
        "required": [
          "checkinId",
          "distanceFromBoxInMeters",
          "consistency",
          "message"
        ],
        "properties": {
          "checkinId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "distanceFromBoxInMeters": {
            "type": "number",
            "example": 42.8
          },
          "consistency": {
            "$ref": "#/components/schemas/CheckinConsistencyResult"
          },
          "message": {
            "type": "string",
            "example": "Check-in realizado com sucesso"
          }
        }
      },
      "CheckinConsistencyResult": {
        "type": "object",
        "required": [
          "activityStatus",
          "currentStreak",
          "longestStreak",
          "availableFreezes",
          "totalXp",
          "xpGained",
          "freezeUsed",
          "milestonesUnlocked"
        ],
        "properties": {
          "activityStatus": {
            "type": "string",
            "enum": [
              "counted",
              "already-counted"
            ]
          },
          "currentStreak": {
            "type": "number",
            "example": 7
          },
          "longestStreak": {
            "type": "number",
            "example": 12
          },
          "availableFreezes": {
            "type": "number",
            "example": 1
          },
          "totalXp": {
            "type": "number",
            "example": 270
          },
          "xpGained": {
            "type": "number",
            "example": 10
          },
          "freezeUsed": {
            "type": "boolean",
            "example": false
          },
          "milestonesUnlocked": {
            "type": "array",
            "items": {
              "type": "number"
            },
            "example": [
              7
            ]
          }
        }
      },
      "Checkin": {
        "type": "object",
        "required": [
          "_id",
          "userId",
          "boxId",
          "latitude",
          "longitude",
          "distanceFromBoxInMeters",
          "createdAt"
        ],
        "properties": {
          "_id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "userId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "boxId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "distanceFromBoxInMeters": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Exercise": {
        "type": "object",
        "required": [
          "_id",
          "name",
          "category",
          "isGlobal",
          "createdAt"
        ],
        "properties": {
          "_id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "name": {
            "type": "string",
            "example": "Snatch"
          },
          "category": {
            "$ref": "#/components/schemas/ExerciseCategory"
          },
          "isGlobal": {
            "type": "boolean"
          },
          "boxId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateExerciseResponse": {
        "type": "object",
        "required": [
          "exerciseId"
        ],
        "properties": {
          "exerciseId": {
            "$ref": "#/components/schemas/ObjectId"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "statusCode": {
            "type": "number",
            "example": 400
          }
        }
      },
      "RegisterUserDto": {
        "type": "object",
        "required": [
          "name",
          "email",
          "password"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120,
            "example": "Joao Aluno"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "joao@aluno.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "Aluno@123"
          }
        }
      },
...
      "NearbyBox": {
        "type": "object",
        "required": [
          "boxId",
          "name",
          "cnpj",
          "latitude",
          "longitude",
          "geofenceRadius",
          "distanceInMeters",
          "isStudentRegistered"
        ],
        "properties": {
          "boxId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "name": {
            "type": "string",
            "example": "Cross Box Alpha"
          },
          "cnpj": {
            "type": "string",
            "example": "12345678000199"
          },
          "latitude": {
            "type": "number",
            "example": -23.56447
          },
          "longitude": {
            "type": "number",
            "example": -46.65284
          },
          "geofenceRadius": {
            "type": "number",
            "example": 100
          },
          "distanceInMeters": {
            "type": "number",
            "example": 42.8
          },
          "isStudentRegistered": {
            "type": "boolean",
            "example": true
          }
        }
      }
    }
  }
}